Rename/move neighbor search files
