+++ parcels/interaction/__init__.py
@@ -1,4 +1,3 @@
 from .baseinteractionkernel import BaseInteractionKernel  # noqa
 from .interactionkernelsoa import InteractionKernelSOA  # noqa
-from .hash_spherical import HashSphericalNeighborSearch  # noqa
-from .brute_force import BruteFlatNeighborSearch  # noqa
+from .neighborsearch import *
