+++ parcels/interaction/neighborsearch/__init__.py
@@ -1,0 +1,5 @@
+from parcels.interaction.neighborsearch.hashflat import HashFlatNeighborSearch
+from parcels.interaction.neighborsearch.hashspherical import HashSphericalNeighborSearch  #noqa
+from parcels.interaction.neighborsearch.bruteforce import BruteFlatNeighborSearch  #noqa
+from parcels.interaction.neighborsearch.bruteforce import BruteSphericalNeighborSearch  #noqa
+from parcels.interaction.neighborsearch.kdtreeflat import KDTreeFlatNeighborSearch  #noqa
