+++ parcels/interaction/neighborsearch/bruteforce.py
@@ -1,10 +1,9 @@
-from parcels.interaction.base_neighbor import BaseFlatNeighborSearch
-from parcels.interaction.base_neighbor import BaseSphericalNeighborSearch
+from parcels.interaction.neighborsearch.base import BaseFlatNeighborSearch
+from parcels.interaction.neighborsearch.base import BaseSphericalNeighborSearch
 
 
 class BruteFlatNeighborSearch(BaseFlatNeighborSearch):
     '''Brute force implementation to find the neighbors.'''
-    name = "brute force"
 
     def find_neighbors_by_coor(self, coor):
         return self._get_close_neighbor_dist(coor, self.active_idx)
@@ -12,7 +11,6 @@ class BruteFlatNeighborSearch(BaseFlatNeighborSearch):
 
 class BruteSphericalNeighborSearch(BaseSphericalNeighborSearch):
     '''Brute force implementation to find the neighbors.'''
-    name = "brute force"
 
     def find_neighbors_by_coor(self, coor):
         return self._get_close_neighbor_dist(coor, self.active_idx)
